+++ src/LionWeb.Core/Core/Notification/Partition/IPartitionNotification.cs
@@ -293,7 +293,7 @@ public record ChildMovedFromOtherContainmentInSameParentNotification(
 /// <param name="Parent"></param>
 /// <param name="Containment"></param>
 /// <param name="OldIndex"></param>
-public record ChildMovedInSameContainmentNotification(
+public record ChildMovedInSameContainmentInSameParentNotification(
     Index NewIndex,
     IWritableNode MovedChild,
     IWritableNode Parent,